Switched from marked to MarkdownIt as a rendering engine.
marked has a security advisory, which nobody knows how to fix,
and there is no timeline for that. Frankly, the security hole is
totally harmless in the setting of this grunt plugin, because no
untrusted code is ever included, all code belongs to the user.
Yet, the security advisory shows up on the "dependencies" badge
upsetting users neeedlessly. So I switched it to markdown-it, and
added quite a few plugins, while I was at it.
Note: markdown-it has a different set of options --- do not forget
to migrate them for your existing presentations.